Inline ConstructorArgumentsShouldMatch test data into tests
